Utilities to install the Corpus Workbench (CWB)

The CWB comprises a set of command line tools for corpus preparation and management. Functionality for installing and managing these Tools.

cwb_install(
  url_cwb = cwb_get_url(),
  md5 = attr(url_cwb, "md5"),
  cwb_dir = fs::path(fs::path_temp(), "cwb"),
  verbose = TRUE
)

cwb_get_url()

cwb_get_bindir(bindir = Sys.getenv("CWB_BINDIR"), verbose = TRUE)

cwb_is_installed()

Arguments

url_cwb

URL for downloading the CWB.

md5

The md5 checksum of the compressed file to be downloaded.

cwb_dir

The directory where the CWB shall be installed.

verbose

Logical, whether to show progress messages.

bindir

The directory with CWB binaries.

Details

Use cwb_install() to download and install CWB binaries (v3.5) from SourceForge. If successful, cwb_install() returns the directory of the CWB, otherwiese NULL. For the installation on macOS and Linux, see https://cwb.sourceforge.io/install.php.

cwb_get_url() will return the URL for downloading the appropriate binary (Linux / macOS) of the CWB (v3.5), or the source tarball (Linux). The md5 checksum of the file to be downloaded is part of the return value as "md5" attribute.

cwb_get_bindir() detects the directory with the cwb command line programs. Defaults to using the value of the environment variable "CWB_BINDIR". If unset, the value of cwb-config --bindir is used. Returns NULL if CWB installation is not found.

cwb_is_installed() will check whether the CWB is installed.