Install and Manage corpora.

Utitlity functions for making the installation of indexed CWB corpora wrapped into R data packages as easy as possible. A data package with a CWB corpus is assumed to include a directory /extdata/cwb/registry for registry files and a directory /extdata/cwb/indexed_corpora for the inexed corpus files.

corpus_install(pkg = NULL, repo = "http://polmine.sowi.uni-due.de/packages",
  tarball = NULL, lib = .libPaths()[1], verbose = TRUE, ...)

corpus_packages()

corpus_copy(old, new, registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  verbose = TRUE)

corpus_rename(old, new, registry_dir = Sys.getenv("CORPUS_REGISTRY"),
  verbose = TRUE)

corpus_remove(corpus, registry_dir = Sys.getenv("CORPUS_REGISTRY"))

corpus_as_tarball(corpus, registry_dir, tarfile, verbose = TRUE)

Arguments

pkg
data package name

repo
URL of the repository

tarball
the URL or local path to a tarball with a CWB indexed corpus

lib
directory for R packages, defaults to .libPaths()[1]; the path may not include a whitespace sign

verbose
logical, whether to be verbose

...
further parameters that will be passed into install.packages

old
name of the (old) corpus

new
name of the (new) corpus

registry_dir
directory of registry

corpus
a CWB corpus

tarfile
filename of tarball

Details

The corpus_install function combines two steps necessary to install a CWB corpus wrapped into a R data package. First, it calls install.packages, then it resets the path pointing to the directory with the indexed corpus files in the registry file. The package will be installed to the standard library directory for installing R packages (.libPaths()[1]). Another location can be used by stating the param 'lib' explicitly (see documentation for install.packages). The function can also be used to install a corpus from a password protected repository. Further parameters are handed over to install.packages, so you might add method = "wget" extra = "--user donald --password duck". See examples how to check whether the directory has been set correctly.

corpus_packages will detect the packages that include CWB corpora.

corpus_copy will create a copy of a corpus (useful for experimental modifications, for instance).

corpus_rename will rename a corpus, affecting the name of the registry file, the corpus id, and the name of the directory where data files reside.

corpus_remove can be used to drop a corpus.

corpus_as_tarball will create a tarball (.tar.gz-file) with two subdirectories. The 'registry' subdirectory will host the registry file for the tarred corpus. The data files will be put in a subdirectory with the corpus name in the 'indexed_corpora' subdirectory.

See also